baidu/requirements.txt
beautifulsoup4==4.6.3
bs4==0.0.1
requests==2.20.1
python-dateutil==2.8.1
hanziconv==0.3.2
